+++ custom_components/tahoma/cover_devices/awning.py
